Use an int rather than a size_t here, I'm sick of including stdint.h on Unix

+++ test/pylTestOverloads.cpp
@@ -3,7 +3,7 @@
 struct Vector4 
 {
 	float x, y, z, w;
-	float& operator[](size_t idx) const { return ((float *)this)[idx]; }
+	float& operator[](int idx) const { return ((float *)this)[idx]; }
 };
 
 // Declare these conversions before including pyliaision.h
